build.sh : Reformat and extend
Add support for multiple build configs and colorize everything

+++ build.sh
@@ -17,7 +17,7 @@ KERNEL="zImage"
 
 #halogen Kernel Details
 KERNEL_NAME="Caesium"
-VER="v1.5"
+VER="v1.6"
 VER="-$(date +"%Y%m%d")-$VER"
 DEVICE="-$(echo $DEFCONFIG | cut -d _ -f 2)"
 FINAL_VER="$KERNEL_NAME""$DEVICE""$VER"
@@ -34,61 +34,73 @@ RESOURCE_DIR="/home/msfjarvis/git-repos/halogenOS/android_kernel_cyanogen_msm891
 ANYKERNEL_DIR="$RESOURCE_DIR/AnyKernel2"
 TOOLCHAIN_DIR="/home/msfjarvis/git-repos/toolchains/arm-eabi-4.8/"
 REPACK_DIR="$ANYKERNEL_DIR"
-#PATCH_DIR="$ANYKERNEL_DIR/patch"
-#MODULES_DIR="$ANYKERNEL_DIR/modules"
 ZIP_MOVE="$RESOURCE_DIR/kernel_out/"
 ZIMAGE_DIR="$KERNEL_DIR/arch/arm/boot"
 
 # Functions
-function make_kernel {
-#xu
-                if [ $1 ];then make clean;fi
-		make $DEFCONFIG $THREAD
-		make $KERNEL $THREAD
-                #make dtbs $THREAD
-		[ -f $ZIMAGE_DIR/$KERNEL ] && cp -vr $ZIMAGE_DIR/$KERNEL $REPACK_DIR/tools/zImage || exit 1
+function make_kernel() {
+  [ $CLEAN ] && make clean
+  make $DEFCONFIG $THREAD
+  make $KERNEL $THREAD
+	[ -f $ZIMAGE_DIR/$KERNEL ] && cp -vr $ZIMAGE_DIR/$KERNEL $REPACK_DIR/tools/zImage || exit 1
 }
 
-#function make_modules {
-#		cd $KERNEL_DIR
-#		make modules $THREAD
-#		find $KERNEL_DIR -name '*.ko' -exec cp {} $MODULES_DIR/ \;
-#		cd $MODULES_DIR
-#        $STRIP --strip-unneeded *.ko
-#        cd $KERNEL_DIR
-#}
-
-function make_dtb {
-		$KERNEL_DIR/dtbToolCM -2 -o $KERNEL_DIR/arch/arm/boot/dt.img -s 2048 -p $KERNEL_DIR/scripts/dtc/ $KERNEL_DIR/arch/arm/boot/dts/
-		cp -vr $KERNEL_DIR/arch/arm/boot/dt.img $REPACK_DIR/tools/dt.img
+function make_dtb() {
+  $KERNEL_DIR/dtbToolCM -2 -o $KERNEL_DIR/arch/arm/boot/dt.img -s 2048 -p $KERNEL_DIR/scripts/dtc/ $KERNEL_DIR/arch/arm/boot/dts/
+  [ -f "$KERNEL_DIR/arch/arm/boot/dt.img" ] && cp -vr $KERNEL_DIR/arch/arm/boot/dt.img $REPACK_DIR/tools/dt.img || exit 1
 }
 
-function make_zip {
-		cd $REPACK_DIR
-        zip -ur kernel_temp.zip *
-        mkdir -p $ZIP_MOVE
-		cp  kernel_temp.zip $ZIP_MOVE/`echo $FINAL_VER`.zip
-		cd $KERNEL_DIR
+function make_zip() {
+  cd $REPACK_DIR
+  zip -ur kernel_temp.zip *
+  mkdir -p $ZIP_MOVE
+  cp  kernel_temp.zip $ZIP_MOVE/$FINAL_VER.zip
+  cd $KERNEL_DIR
 }
+
+function tg() {
+  curl -F chat_id="$TG_ID" -F document="@$1" "https://api.telegram.org/bot$TG_BOT_ID/sendDocument"
+}
+
+function upload() {
+  echo -e "${cyan} Uploading file to Telegram ${restore}"
+  tg kernel_out/$FINAL_VER.zip
+}
+
+while getopts ":ctu" opt; do
+  case $opt in
+    c)
+      echo -e "${cyan} Building clean ${restore}" >&2
+      CLEAN=true
+      ;;
+    t)
+      echo -e "${cyan} Will upload build to Telegram! ${restore}" >&2
+      UPLOAD=true
+      ;;
+    u)
+      echo -e "${cyan} Only making ZIP and uploading to Telegram! ${restore}" >&2
+      ONLY_UPLOAD=true
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      ;;
+  esac
+done
 
 DATE_START=$(date +"%s")
 
 # TC tasks
 export CROSS_COMPILE=$TOOLCHAIN_DIR/bin/arm-eabi-
 export LD_LIBRARY_PATH=$TOOLCHAIN_DIR/lib/
-#STRIP=$TOOLCHAIN_DIR/uber-4.9/bin/aarch64-linux-android-strip
-#rm -rf $MODULES_DIR/*
-rm -rf $ZIP_MOVE/*
-#rm -rf $KERNEL_DIR/arch/arm/boot/dt.img
 cd $ANYKERNEL_DIR/tools
 rm -rf $ZIMAGE_DIR/zImage
 rm -rf dt.img
 cd $KERNEL_DIR
 
 # Make
+[ $ONLY_UPLOAD ] && make_zip && upload && exit 0
 make_kernel
 make_dtb
-#make_modules
 make_zip
 
 echo -e "${green}"
@@ -100,3 +112,5 @@ DATE_END=$(date +"%s")
 DIFF=$(($DATE_END - $DATE_START))
 echo "Time: $(($DIFF / 60)) minute(s) and $(($DIFF % 60)) seconds."
 echo " "
+
+[ $UPLOAD ] && upload
